testing a refactor of tableau-optimize
